requirements.txt
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
deep-translator==1.11.4
Django==5.0.6
django-scheduler==0.10.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
icalendar==5.0.12
idna==3.10
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
pyaes==1.6.1
pyasn1==0.5.1
pyTelegramBotAPI==4.17.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==21.11.1
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.0
telebot==0.0.5
Telethon==1.34.0
tqdm==4.66.1
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.3.0
